-#142: ordenacion de los subsitemas en la ficha tecnica

+++ sis_mantenimiento/base/funciones/gem.f_uni_cons_det_sel.sql
@@ -68,7 +68,7 @@ BEGIN
             ucdet.id_uni_cons_det,
             ucdet.id_unidad_medida,
             ucdet.id_uni_cons,
-            ucdet.codigo,
+            ucdet.codigo::numeric,
             ucdet.nombre,
             ucdet.descripcion,
             ucdet.valor,
